Extract resolveUrl methods
